https://myddleware.atlassian.net/browse/PROJ-647 move confirmation to delete the rule to dedicated rule detail view
