+++ assets/app.js
@@ -45,6 +45,10 @@ if (window.location.href.includes('rule/list')) {
     require('./js/rule-list.js');
 }
 
+if (window.location.href.includes('rule/view')) {
+    require('./js/rule-detail.js');
+}
+
 if (window.location.href.includes('rule/document/list')) {
     require('./js/filter.js');
 }
